requirements.txt
pdfplumber>=0.11.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
